backend/requirements.txt
django==5.2
djangorestframework==3.16.0
djangorestframework-simplejwt==5.5.0
django-cors-headers==4.7.0
django-filter==25.1
pytz==2024.1 